#2 Refactor: Move all Rule classes to new com.example.linter.rule package
