+++ src/main/java/com/example/linter/config/blocks/AbstractBlock.java
@@ -1,8 +1,8 @@
 package com.example.linter.config.blocks;
 
 import com.example.linter.config.BlockType;
-import com.example.linter.config.OccurrenceRule;
 import com.example.linter.config.Severity;
+import com.example.linter.rule.OccurrenceRule;
 
 import java.util.Objects;
