+++ src/main/java/com/example/linter/rule/SectionRule.java
@@ -1,4 +1,4 @@
-package com.example.linter.config;
+package com.example.linter.rule;
 
 import com.example.linter.config.blocks.AbstractBlock;
