+++ src/main/java/com/example/linter/rule/TitleRule.java
@@ -1,4 +1,4 @@
-package com.example.linter.config;
+package com.example.linter.rule;
 
 import java.util.Objects;
